Provide known defaults during upstream install.
The defaults names for the initial organization and location generated
with the `Katello` installer (upstream) are named `Default Organization`
and `Default Location` (note the space between words), but downstream
the names have an underscore in them. By enforcing the use of an
underscore in their names during a `Katello` installation, we can ensure
that we can search for these entities by name.