Switched from yum-config-manager to sub-manager.
No longer installing `yum-config-manager` for ``Satellite 6`` installs
as it can take a while for YUM to re-fresh its cache. Instead, use
`subscription-manager repos` to first disable everything and then
enabled the RHEL repos we want. Note that since our custom `satellite`
repo is not a RHEL repo and is enabled by default, we do not have to
explicitly enable it anymore.

This should speed up the installs considerably.