Moved system registration to own method.
Since we currently have 2 different tasks that use the same code to
register and subscribe a host to Red Hat, I have moved this code into a
new `subscribe` method and refactored the code. I have also added a new
`unsubscribe` method that can be used to `clean up` a host.