Create a task specific for the provisioning job
Rename task reservation_install to provision_install in order to
transform it to be specific to the provisioning job.

Improve the task to be able to clean the VM before each installation.
And add option to install from an ISO image.

Also make iso_install task atomic. Subscribe and install_prerequisites
should be done outside the iso_install task.