Refactor product installer
Refactor the product_install task to avoid calling the katello-installer
several times. All katello-installer options will be aggregated to run
the katello-installer command at the very end, just before setting the
fake manifest.

Add an argument to the tasks which calls katello-installer to define if
the katello-installer should be run or not. If not, then the task will
return the options it should be using on the katello-installer.

Clean the setup_proxy task, now just the options on the
katello-installer are enough.