Capsule provisioning
Add setup_capsule_firewall task and update the setup_firewall to
setup_satellite_firewall. The setup_firewall task now is a helper for
setting up firewall rules and using iptables on RHEL6 and firewall-cmd
on RHEL7.

Add setup_capsule_content task to help setting up a custom content on
the Satellite Server that allow creating a capsule.

Update capsule_installer to match the new way to run the capsule
installer.

Update setup_capsules to not start content sync. This may be added later
if syncing is something really needed.