Update automation-tools to separate out target_image and hostname
a) This is required because if target_image and hostname remain same
then we would have the hostname for provisioning job as,
`qe-sat6-rhel7-base.satqe.lab.eng.rdu2.redhat.com`, which we do not
as it will cause DNS issues.
b) So separating target_image and hostname would enable us to have,
the hostname as `qe-sat6-rhel7.satqe.lab.eng.rdu2.redhat.com` and
the provisioning job could creae a virsh domain or image name with
`qe-sat6-rhel7-base`.